Moved .ui files to a separate directory and refactored out supporting functions into utils.py
