+++ lyse/main.py
@@ -53,6 +53,7 @@ import qtutils.icons
 # Lyse imports
 from lyse.dataframe_utilities import concat_with_padding, get_dataframe_from_shot, replace_with_padding
 from lyse import LYSE_DIR, _rangeindex_to_multiindex
+import lyse.utils
 
 @inmain_decorator()
 def error_dialog(app, message):
@@ -64,88 +65,6 @@ def question_dialog(app, message):
     reply = QtWidgets.QMessageBox.question(app.ui, 'lyse', message,
                                        QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No)
     return (reply == QtWidgets.QMessageBox.Yes)
-
-
-def scientific_notation(x, sigfigs=4, mode='eng'):
-    """Returns a unicode string of the float f in scientific notation"""
-
-    times = u'\u00d7'
-    thinspace = u'\u2009'
-    hairspace = u'\u200a'
-    sups = {u'-': u'\u207b',
-            u'0': u'\u2070',
-            u'1': u'\xb9',
-            u'2': u'\xb2',
-            u'3': u'\xb3',
-            u'4': u'\u2074',
-            u'5': u'\u2075',
-            u'6': u'\u2076',
-            u'7': u'\u2077',
-            u'8': u'\u2078',
-            u'9': u'\u2079'}
-
-    prefixes = {
-        -24: u"y",
-        -21: u"z",
-        -18: u"a",
-        -15: u"f",
-        -12: u"p",
-        -9: u"n",
-        -6: u"\u03bc",
-        -3: u"m",
-        0: u"",
-        3: u"k",
-        6: u"M",
-        9: u"G",
-        12: u"T",
-        15: u"P",
-        18: u"E",
-        21: u"Z",
-        24: u"Y"
-    }
-
-    if not isinstance(x, float):
-        raise TypeError('x must be floating point number')
-    if np.isnan(x) or np.isinf(x):
-        return str(x)
-    if x != 0:
-        exponent = int(np.floor(np.log10(np.abs(x))))
-        # Only multiples of 10^3
-        exponent = int(np.floor(exponent / 3) * 3)
-    else:
-        exponent = 0
-
-    significand = x / 10 ** exponent
-    pre_decimal, post_decimal = divmod(significand, 1)
-    digits = sigfigs - len(str(int(pre_decimal)))
-    significand = round(significand, digits)
-    result = str(significand)
-    if exponent:
-        if mode == 'exponential':
-            superscript = ''.join(sups.get(char, char) for char in str(exponent))
-            result += thinspace + times + thinspace + '10' + superscript
-        elif mode == 'eng':
-            try:
-                # If our number has an SI prefix then use it
-                prefix = prefixes[exponent]
-                result += hairspace + prefix
-            except KeyError:
-                # Otherwise display in scientific notation
-                superscript = ''.join(sups.get(char, char) for char in str(exponent))
-                result += thinspace + times + thinspace + '10' + superscript
-    return result
-
-
-def get_screen_geometry(qapplication):
-    """Return the a list of the geometries of each screen: each a tuple of
-    left, top, width and height"""
-    geoms = []
-    desktop = qapplication.desktop()
-    for i in range(desktop.screenCount()):
-        sg = desktop.screenGeometry(i)
-        geoms.append((sg.left(), sg.top(), sg.width(), sg.height()))
-    return geoms
-
 
 class WebServer(ZMQServer):
 
@@ -467,7 +386,7 @@ class RoutineBox(object):
         
         loader = UiLoader()
         loader.registerCustomWidget(TreeView)
-        self.ui = loader.load(os.path.join(LYSE_DIR, 'routinebox.ui'))
+        self.ui = loader.load(os.path.join(LYSE_DIR, 'user_interface/routinebox.ui'))
         container.addWidget(self.ui)
 
         if multishot:
@@ -839,7 +758,7 @@ class EditColumns(object):
         self.old_columns_visible = columns_visible.copy()
 
         loader = UiLoader()
-        self.ui = loader.load(os.path.join(LYSE_DIR, 'edit_columns.ui'), EditColumnsDialog())
+        self.ui = loader.load(os.path.join(LYSE_DIR, 'user_interface/edit_columns.ui'), EditColumnsDialog())
 
         self.model = UneditableModel()
         self.header = HorizontalHeaderViewWithWidgets(self.model)
@@ -1450,7 +1369,7 @@ class DataFrameModel(QtCore.QObject):
                     continue
             value = dataframe_row[column_name]
             if isinstance(value, float):
-                value_str = scientific_notation(value)
+                value_str = lyse.utils.scientific_notation(value)
             else:
                 value_str = str(value)
             lines = value_str.splitlines()
@@ -1610,7 +1529,7 @@ class FileBox(object):
 
         loader = UiLoader()
         loader.registerCustomWidget(TableView)
-        self.ui = loader.load(os.path.join(LYSE_DIR, 'filebox.ui'))
+        self.ui = loader.load(os.path.join(LYSE_DIR, 'user_interface/filebox.ui'))
         self.ui.progressBar_add_shots.hide()
         container.addWidget(self.ui)
         self.shots_model = DataFrameModel(self.app, self.ui.tableView, self.exp_config)
@@ -1939,7 +1858,7 @@ class Lyse(object):
     def __init__(self, qapplication):
         self.qapplication = qapplication
         loader = UiLoader()
-        self.ui = loader.load(os.path.join(LYSE_DIR, 'main.ui'), LyseMainWindow(self))
+        self.ui = loader.load(os.path.join(LYSE_DIR, 'user_interface/main.ui'), LyseMainWindow(self))
 
         self.process_tree = ProcessTree.instance()
 
@@ -2139,7 +2058,7 @@ class Lyse(object):
 
         save_data['window_pos'] = (window_pos.x(), window_pos.y())
 
-        save_data['screen_geometry'] = get_screen_geometry(self.qapplication)
+        save_data['screen_geometry'] = lyse.utils.get_screen_geometry(self.qapplication)
         save_data['splitter'] = self.ui.splitter.sizes()
         save_data['splitter_vertical'] = self.ui.splitter_vertical.sizes()
         save_data['splitter_horizontal'] = self.ui.splitter_horizontal.sizes()
@@ -2222,7 +2141,7 @@ class Lyse(object):
         # position was saved when 2 monitors were plugged in but there is
         # only one now, and the splitters may not make sense in light of a
         # different window size, so better to fall back to defaults:
-        current_screen_geometry = get_screen_geometry(self.qapplication)
+        current_screen_geometry = lyse.utils.get_screen_geometry(self.qapplication)
         if current_screen_geometry == screen_geometry:
             if 'window_size' in save_data:
                 self.ui.resize(*save_data['window_size'])
