+++ lyse/user_interface/plot_window.ui
